Refactor Task into Job
Refactor Task Class into Job Class because they served similar functions.
