+++ src/applications/MachineShopSimulator.java
@@ -38,7 +38,7 @@ public class MachineShopSimulator {
     }
 
     private int getMachineForNextTask(Job theJob) {
-        int p = ((Task) theJob.getTaskQ().getFrontElement()).getMachine();
+        int p = ((Job) theJob.getTaskQ().getFrontElement()).getMachine();
         return p;
     }
 
@@ -89,7 +89,7 @@ public class MachineShopSimulator {
     private void setUpJobs(SimulationSpecification specification) {
         Job theJob;
         for (int i = 1; i <= specification.getNumJobs(); i++) {
-            int tasks = Task.getNumTasks(specification, i);
+            int tasks = Job.getNumTasks(specification, i);
             int firstMachine = 0; // machine for first task
 
             theJob = new Job(i);
