--- src/applications/Task.java
@@ -1,25 +1,0 @@
-package applications;
-
-// top-level nested classes
-class Task {
-    // data members
-    private int machine;
-    private int time;
-
-    // constructor
-    Task(int theMachine, int theTime) {
-        machine = theMachine;
-        time = theTime;
-    }
-
-    public int getMachine() {
-        return machine;
-    }
-
-    public int getTime() {
-        return time;
-    }
-    public static int getNumTasks(SimulationSpecification specification, int i) {
-        return specification.getJobSpecifications(i).getNumTasks();
-    }
-}
